// app/clinix-ai-billing/embedded/types.ts
export enum ViewState {
  DASHBOARD = 'DASHBOARD',
  PRIORITIZATION = 'PRIORITIZATION',
  PERFORMANCE_REVIEW = 'PERFORMANCE_REVIEW',
  DENIAL_MANAGER = 'DENIAL_MANAGER',
  CLAIM_METHOD = 'CLAIM_METHOD',
  CLAIM_WIZARD = 'CLAIM_WIZARD',
  UPLOAD_REVIEW = 'UPLOAD_REVIEW',
  CLAIM_DETAIL = 'CLAIM_DETAIL',
  SETTINGS = 'SETTINGS',
  SUCCESS = 'SUCCESS'
}

export interface NavItem {
  label: string;
  view: ViewState;
  icon?: string;
}